fix(rust/driver_manager): remove optional property for windows deps (#3693)
Hi there! When using the `adbc_driver_manager` crate on Windows, it's
impossible to build the project because the `adbc_driver_manager`
dependency fails to compile. Windows crates are marked as optional and
are not included in any of the crate's features. To make it work, the
`optional` flag needs to be removed.

Signed-off-by: Pavel Agafonov <pavel.agafonov.al@gmail.com>