fix(go/adbc/sqldriver): properly close resources (#3731)
fixes #3728

This was a combination of two issues that are fixed in this PR:

1. `sqldriver.Connector` needs to implement the `io.Closer` interface so
that `sql.DB.Close` will call it and call `Close` on the `adbc.Database`
object.
2. `rows.Close` needs to also close the associated `stmt`. In the case
of at least one driver (the duckdb driver in associated issue) not
closing the `stmt` led to the handle to the duckdb database not being
dropped despite calling `Close` on the database.