fix(c): Generate versioned DLLs and import LIBs when building with MSVC (#2858)
Hi.

This is to address building binaries and LIBs on Windows. The PR is
mainly for consideration purposes, as it's my first PR against this repo
so not necessarily a complete one. Following are the changes I needed to
make:
- Disable ASAN/UBSAN builds configs as they require GCC/Clang, and on
Windows we're building with MSVC.
- An environment variable to `vcpkg` root is expected so that third
party packages (sqlite, postgres etc) can be installed and found via
vcpkg.
- Disable some warnings for testing code that are safe to ignore for
MSVC builds.
- Ensure the generated binaries have file level details (Details tab on
the file properties in Windows) for the version. This is required by
Windows MSIs that may be used to install the adbc DLLs in order to
detect if the DLL being installed is a newer or older version.
- Generate import LIB file for drivers built with go build, which
includes flight SQL and Snowflake drivers. This allows the testing
projects referencing those DLLs to built on Windows.
- The `GoUtils.cmake` is not automatically copying the driver DLLs to
the location of the test executables, causing the tests for those
drivers to fail. Added a post build command to ensure the DLLs are
copied.
- Fix some type casting mismatch (casting between size_t and int).
- Fix warnings where a variable is hiding a previous one with the same
name. Appended `2` to the second variable name.

What I haven't done is update the GitHub actions to verify a complete
build on a Windows machine as I'm not too familiar with the CI/CD
pipeline in this repo.

Hopefully this is a good start to get everything building successfully
and Windows.

Closes #2846

---------

Co-authored-by: Sutou Kouhei <kou@cozmixng.org>
Co-authored-by: David Li <li.davidm96@gmail.com>
Co-authored-by: Bryce Mecum <petridish@gmail.com>